Almost Done
I Tried doing the carousel without libraries but my efforts wasn't good enough.
Main limitation was the looping and the touch accessibility
But thanks to UI kit and it was done in a jiffy.
For the on-view animations i used some already prepared on stylesheet, then just and the custom animation attributes to the preferred element